Added Author and Genre features, implemented and fully tested correctly.

+++ src/controllers/helper.js
@@ -1,15 +1,20 @@
-const { Book, Reader } = require('../models');
+const { Book, Reader, Author, Genre } = require('../models');
 
 const getModel = (model) => {
     const models = {
         reader: Reader,
-        book: Book
+        book: Book,
+        author: Author,
+        genre: Genre
     };
     return models[model];
 }
 
 const readerError = { error: 'The reader could not be found.' }
 const bookError = { error: 'The book could not be found.' }
+const authorError = { error: 'The author could not be found.' }
+const genreError = { error: 'The genre could not be found.' }
+
 
 const removePassword = (obj) => {
     if (obj.hasOwnProperty('password')) {
@@ -19,7 +24,7 @@ const removePassword = (obj) => {
 };
 
 const create = async (model, req, res) => {
-    const thisItem = await getModel(model).create(req.body)
+    const thisItem = await getModel(model).create(req.body);
     res.status(201).send(removePassword(thisItem.dataValues));
 }
 
@@ -39,6 +44,12 @@ const findById = async (model, req, res) => {
     if (!thisItem && model == 'book') {
         return res.status(404).send(bookError);
     }
+    if (!thisItem && model == 'author') {
+        return res.status(404).send(authorError);
+    }
+    if (!thisItem && model == 'genre') {
+        return res.status(404).send(genreError);
+    }
     res.status(200).send(removePassword(thisItem.dataValues));
 }
 
@@ -49,6 +60,12 @@ const update = async (model, req, res) => {
     }
     if (!thisItem && model == 'book') {
         return res.status(404).send(bookError);
+    }
+    if (!thisItem && model == 'author') {
+        return res.status(404).send(authorError);
+    }
+    if (!thisItem && model == 'genre') {
+        return res.status(404).send(genreError);
     }
     await getModel(model).update(req.body, {
         where: { id: req.params.id },
@@ -65,6 +82,12 @@ const remove = async (model, req, res) => {
     }
     if (!thisItem && model == 'book') {
         return res.status(404).send(bookError);
+    }
+    if (!thisItem && model == 'author') {
+        return res.status(404).send(authorError);
+    }
+    if (!thisItem && model == 'genre') {
+        return res.status(404).send(genreError);
     }
     await getModel(model).destroy({
         where: { id: req.params.id }
